chore: migrate tests to typescript
