+++ frontend/automation/e2e/settings.cy.ts
