+++ frontend/automation/e2e/wallet-remove.cy.ts
@@ -1,8 +1,8 @@
-const { unlockWallet } = require('../support/helpers')
+import { unlockWallet } from '../support/helpers'
 
 describe('wallet remove', () => {
-  let walletName
-  let passphrase
+  let walletName: string
+  let passphrase: string
   const form = 'remove-wallet-form'
 
   before(() => {
@@ -27,25 +27,21 @@ describe('wallet remove', () => {
     // 0001-WALL-068 must be able to remove a wallet
     unlockWallet(walletName, passphrase)
     cy.getByTestId('remove-wallet').click()
-    submitForm()
+    submitForm(form)
     cy.getByTestId('helper-text').contains('Required')
     cy.getByTestId(form).get('input[type="text"]').type('invalid text')
-    submitForm()
+    submitForm(form)
     cy.getByTestId('helper-text').contains('Invalid confirmation text')
     cy.getByTestId(form)
       .get('input[type="text"]')
       .clear()
       .type(`Remove ${walletName}`)
-    submitForm()
+    submitForm(form)
     cy.getByTestId('toast').should('contain.text', 'Wallet removed')
     cy.getByTestId('wallet-home').should('be.visible')
     cy.getByTestId(`wallet-${walletName}`).should('not.exist')
   })
-
-  function submitForm() {
-    cy.getByTestId(form)
-      .should('be.visible')
-      .get('button[type="submit"]')
-      .click()
-  }
 })
+function submitForm(form: string) {
+  cy.getByTestId(form).should('be.visible').get('button[type="submit"]').click()
+}
