+++ frontend/automation/e2e/wallet-service.cy.ts
