+++ frontend/automation/support/commands.ts
@@ -1,0 +1,13 @@
+import './commands/get-by-test-id'
+import './commands/clean'
+import './commands/backend'
+import './commands/set-vega-home'
+import './commands/restore-wallet'
+import './commands/restore-network'
+import './commands/send-transaction'
+import './commands/send-connection-request'
+import './commands/send-permission-request'
+import './commands/mock-requests'
+import './commands/wait-for-home'
+import './commands/monitor-clipboard'
+import './commands/get-copied-text-from-clipboard'
