--- frontend/automation/support/graphql.js
@@ -1,6 +1,0 @@
-export const hasOperationName = (req, operationName) => {
-  const { body } = req
-  return (
-    body.hasOwnProperty('operationName') && body.operationName === operationName
-  )
-}
