+++ frontend/automation/support/graphql.ts
@@ -1,0 +1,9 @@
+import type { CyHttpMessages } from 'cypress/types/net-stubbing'
+
+export const hasOperationName = (
+  req: CyHttpMessages.IncomingHttpRequest,
+  operationName: string
+) => {
+  const { body } = req
+  return 'operationName' in body && body.operationName === operationName
+}
