+++ frontend/automation/tsconfig.json
@@ -1,10 +1,11 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "allowJs": true,
+    "isolatedModules": false,
+    "allowJs": false,
     "jsx": "react-jsx",
     "types": ["cypress", "node"],
     "sourceMap": false
   },
-  "include": ["src/**/*.ts", "src/**/*.js"]
+  "include": ["**/*.ts", "**/*.js"]
 }
