+++ frontend/package.json
@@ -13,8 +13,8 @@
     "dev": "VITE_ENV='development' && npx vite --port=3000",
     "e2e:clean": "rm -r automation/test-wallets; mkdir automation/test-wallets",
     "e2e:network": "mkdir -p automation/network-config && cp automation/data/networks/test.toml $_",
-    "e2e:open": "yarn run cypress open --config-file automation/cypress.config.js",
-    "e2e:run": "yarn run cypress run --config-file automation/cypress.config.js",
+    "e2e:open": "yarn run cypress open --config-file automation/cypress.config.ts",
+    "e2e:run": "yarn run cypress run --config-file automation/cypress.config.ts",
     "format:check": "npx prettier --check .",
     "format:fix": "npx prettier --write .",
     "lint:check": "eslint ./src",
@@ -40,7 +40,7 @@
     "@ethersproject/providers": "5.7.2",
     "@ethersproject/wallet": "5.7.0",
     "@types/lodash": "4.14.189",
-    "@types/node": "*",
+    "@types/node": "^18.14.0",
     "@types/react-dom": "^18.0.9",
     "@vitejs/plugin-react": "2.2.0",
     "axios": "^1.3.2",
