+++ frontend/tsconfig.json
@@ -17,5 +17,5 @@
     "jsx": "react-jsx"
   },
   "include": ["src"],
-  "exclude": ["node_modules", "src/wailsjs", "automation"]
+  "exclude": ["node_modules", "src/wailsjs"]
 }
